Remove UI behaviour logic from core logic.
