+++ src/test/java/RouterTest.java
@@ -2,12 +2,9 @@ import com.despegar.http.client.GetMethod;
 import com.despegar.http.client.HttpClientException;
 import com.despegar.http.client.HttpResponse;
 import com.despegar.sparkjava.test.SparkServer;
-import game.Board;
 import org.junit.ClassRule;
 import org.junit.Test;
 import spark.servlet.SparkApplication;
-import static game.Mark.*;
-import static java.util.Arrays.asList;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
@@ -34,23 +31,5 @@ public class RouterTest {
         GetMethod request = testServer.get("/", false);
         HttpResponse httpResponse = testServer.execute(request);
         assertTrue((new String(httpResponse.body()).matches(".*1.*2.*3.*4.*5.*6.*7.*8.*9.*")));
-    }
-
-    @Test
-    public void itFormatsQueryStringFromEmptyGridState() {
-        Router router = new Router();
-        Board board = new Board();
-        board.grid = asList(EMPTY, EMPTY, EMPTY);
-
-        assertEquals("123", router.createQueryValueForGridState(board.grid));
-    }
-
-    @Test
-    public void itFormatsQueryStringFromGridStateWithMarks() {
-        Router router = new Router();
-        Board board = new Board();
-        board.grid = asList(EMPTY, X, O);
-
-        assertEquals("1XO", router.createQueryValueForGridState(board.grid));
     }
 }
