+++ src/test/java/game/BoardTest.java
@@ -208,11 +208,4 @@ public class BoardTest {
 
         assertEquals("Tie", board.findResult().getResult());
     }
-
-    @Test
-    public void createsNewBoardFromAString() {
-        Board board = new Board(3, "12X");
-
-        assertEquals(board.grid, asList(EMPTY, EMPTY, X));
-    }
 }
